server/pyproject.toml
[project]
name = "smilex-cas-server"
version = "1.0.0"
description = "SmileX-CAS content management and multi-platform publishing system"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "email-validator>=2.1.0",
    "playwright>=1.40.0",
    "httpx>=0.27.0",
    "aiofiles>=23.0.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
]
